Compute frontiers after integration

+++ se_voxel_impl/src/MultiresOFusion/MultiresOFusion_mapping.cpp
@@ -78,10 +78,7 @@ struct MultiresOFusionUpdate {
                         const float                        voxel_dim,
                         const size_t                       voxel_depth,
                         const float                        max_depth_value,
-                        const unsigned                     frame,
-                        std::set<se::key_t>*               free_nodes,
-                        std::set<se::key_t>*               added_frontier_blocks,
-                        std::set<se::key_t>*               removed_frontier_blocks) :
+                        const unsigned                     frame) :
       map_(map),
       pool_(map.pool()),
       block_list_(block_list),
@@ -100,24 +97,12 @@ struct MultiresOFusionUpdate {
       voxel_depth_(voxel_depth),
       max_depth_value_(max_depth_value),
       frame_(frame),
-      free_nodes_(free_nodes),
-      added_frontier_blocks_(added_frontier_blocks),
-      removed_frontier_blocks_(removed_frontier_blocks),
       zero_depth_band_(1.0e-6f),
       size_to_radius_(std::sqrt(3.0f) / 2.0f) {
 
     corner_rel_steps_ << 0, 1, 0, 1, 0, 1, 0, 1,
         0, 0, 1, 1, 0, 0, 1, 1,
         0, 0, 0, 0, 1, 1, 1, 1;
-    if (free_nodes_) {
-      free_nodes_->clear();
-    }
-    if (added_frontier_blocks_) {
-      added_frontier_blocks_->clear();
-    }
-    if (removed_frontier_blocks_) {
-      removed_frontier_blocks_->clear();
-    }
   };
 
   OctreeType& map_;
@@ -138,9 +123,6 @@ struct MultiresOFusionUpdate {
   const size_t voxel_depth_;
   const float max_depth_value_;
   const unsigned frame_;
-  std::set<se::key_t>* free_nodes_;
-  std::set<se::key_t>* added_frontier_blocks_;
-  std::set<se::key_t>* removed_frontier_blocks_;
   Eigen::Matrix<float, 3, 8> corner_rel_steps_;
   const float zero_depth_band_;
   const float size_to_radius_;
@@ -220,6 +202,7 @@ struct MultiresOFusionUpdate {
                       buffer_data.g = parent_data.g;
                       buffer_data.b = parent_data.b;
                       buffer_data.observed = false; ///<< Set falls such that the observe count can work properly
+                      buffer_data.frontier = parent_data.frontier; // TODO SEM does this make sense?
 
                     } // i
                   } // j
@@ -253,52 +236,6 @@ struct MultiresOFusionUpdate {
       block->incrBufferIntegrCount();
 
       if(block->switchData()) {
-
-//        const unsigned int recommended_stride = 1 << recommended_scale;
-//        const Eigen::Vector3i voxel_coord_base = block->coordinates();
-//
-//        /**
-//         * How would you remove the frontiers of the previous scale???
-//         */
-//
-//        if (added_frontier_blocks_ && removed_frontier_blocks_) {
-//          for (unsigned int z = 0; z < size_at_recommended_scale_li; z++) {
-//            for (unsigned int y = 0; y < size_at_recommended_scale_li; y++) {
-//              for (unsigned int x = 0; x < size_at_recommended_scale_li; x++) {
-//
-//                const int buffer_idx = x + y * size_at_recommended_scale_li + z * size_at_recommended_scale_sq;
-//                auto& buffer_data = block->bufferData(buffer_idx); /// \note pass by reference now.
-//                const VoxelState prev_state =  VoxelState::Unknown; ///< Would this be correct???
-//                // Set the voxel state
-//
-//                if (MultiresOFusion::VoxelType::isFree(buffer_data)) {
-//                  const Eigen::Vector3i coord = voxel_coord_base + recommended_stride * Eigen::Vector3i(x, y, z);
-//                  bool is_frontier;
-//#pragma omp critical (IntegrationCheckFrontier)
-//                  is_frontier = se::isFrontier(coord, size_at_recommended_scale_li, recommended_scale, *block, map_);
-//                  if (is_frontier) {
-//                    buffer_data.state = VoxelState::Frontier;
-//                  } else {
-//                    buffer_data.state = VoxelState::Free;
-//                  }
-//                } else if (MultiresOFusion::VoxelType::isInside(buffer_data)) {
-//                  buffer_data.state = VoxelState::Occupied;
-//                } else {
-//                  buffer_data.state = VoxelState::Unknown;
-//                }
-//                // Update the frontier block sets
-//                if (buffer_data.state == VoxelState::Frontier) {
-//#pragma omp critical (IntegrationAddFrontier)
-//                  added_frontier_blocks_->insert(block->code());
-//                } else if (prev_state == VoxelState::Frontier) {
-//#pragma omp critical (IntegrationRemoveFrontier)
-//                  removed_frontier_blocks_->insert(block->code());
-//                }
-//              }
-//            }
-//          }
-//        }
-
         return;
       }
 
@@ -310,9 +247,6 @@ struct MultiresOFusionUpdate {
     const unsigned int size_at_integration_scale_li = block->size_li >> integration_scale;
     const unsigned int size_at_integration_scale_sq = se::math::sq(size_at_integration_scale_li);
 
-    const unsigned int integration_stride = 1 << integration_scale;
-    const Eigen::Vector3i voxel_coord_base = block->coordinates();
-
     for (unsigned int z = 0; z < size_at_integration_scale_li; z++) {
       for (unsigned int y = 0; y < size_at_integration_scale_li; y++) {
 
@@ -322,36 +256,7 @@ struct MultiresOFusionUpdate {
           // Update the voxel data based using the depth measurement
           const int voxel_idx = x + y * size_at_integration_scale_li + z * size_at_integration_scale_sq;
           auto& voxel_data = block->currData(voxel_idx); /// \note pass by reference now.
-          const VoxelState prev_state = voxel_data.state;
           block->incrCurrObservedCount(updating_model::freeVoxel(voxel_data));
-
-          // Set the voxel state
-          if (added_frontier_blocks_ && removed_frontier_blocks_) {
-            if (MultiresOFusion::VoxelType::isFree(voxel_data)) {
-              const Eigen::Vector3i coord = voxel_coord_base + integration_stride * Eigen::Vector3i(x, y, z);
-              bool is_frontier;
-#pragma omp critical (IntegrationCheckFrontier)
-              is_frontier = se::isFrontier(coord, size_at_integration_scale_li, integration_scale, *block, map_);
-              if (is_frontier) {
-                voxel_data.state = VoxelState::Frontier;
-              } else {
-                voxel_data.state = VoxelState::Free;
-              }
-            } else if (MultiresOFusion::VoxelType::isInside(voxel_data)) {
-              voxel_data.state = VoxelState::Occupied;
-            } else {
-              voxel_data.state = VoxelState::Unknown;
-            }
-            // Update the frontier block sets
-            if (voxel_data.state == VoxelState::Frontier) {
-#pragma omp critical (IntegrationAddFrontier)
-              added_frontier_blocks_->insert(block->code());
-            } else if (prev_state == VoxelState::Frontier) {
-#pragma omp critical (IntegrationRemoveFrontier)
-              removed_frontier_blocks_->insert(block->code());
-            }
-          }
-
         } // x
       } // y
     } // z
@@ -452,6 +357,7 @@ struct MultiresOFusionUpdate {
                       buffer_data.g = parent_data.g;
                       buffer_data.b = parent_data.b;
                       buffer_data.observed = false; ///<< Set falls such that the observe count can work properly
+                      buffer_data.frontier = parent_data.frontier; // TODO SEM does this make sense?
 
                     } // i
                   } // j
@@ -515,49 +421,6 @@ struct MultiresOFusionUpdate {
       block->incrBufferIntegrCount(project_inside);
 
       if(block->switchData()) {
-
-//        /**
-//         * How would you remove the frontiers of the previous scale???
-//         */
-//
-//        if (added_frontier_blocks_ && removed_frontier_blocks_) {
-//          for (unsigned int z = 0; z < size_at_recommended_scale_li; z++) {
-//            for (unsigned int y = 0; y < size_at_recommended_scale_li; y++) {
-//              for (unsigned int x = 0; x < size_at_recommended_scale_li; x++) {
-//
-//                const int buffer_idx = x + y * size_at_recommended_scale_li + z * size_at_recommended_scale_sq;
-//                auto& buffer_data = block->bufferData(buffer_idx); /// \note pass by reference now.
-//                const VoxelState prev_state = VoxelState::Unknown; ///< Would this be correct???
-//                // Set the voxel state
-//
-//                if (MultiresOFusion::VoxelType::isFree(buffer_data)) {
-//                  const Eigen::Vector3i coord = voxel_coord_base + recommended_stride * Eigen::Vector3i(x, y, z);
-//                  bool is_frontier;
-//  #pragma omp critical (IntegrationCheckFrontier)
-//                  is_frontier = se::isFrontier(coord, size_at_recommended_scale_li, recommended_scale, *block, map_);
-//                  if (is_frontier) {
-//                    buffer_data.state = VoxelState::Frontier;
-//                  } else {
-//                    buffer_data.state = VoxelState::Free;
-//                  }
-//                } else if (MultiresOFusion::VoxelType::isInside(buffer_data)) {
-//                  buffer_data.state = VoxelState::Occupied;
-//                } else {
-//                  buffer_data.state = VoxelState::Unknown;
-//                }
-//                // Update the frontier block sets
-//                if (buffer_data.state == VoxelState::Frontier) {
-//  #pragma omp critical (IntegrationAddFrontier)
-//                  added_frontier_blocks_->insert(block->code());
-//                } else if (prev_state == VoxelState::Frontier) {
-//  #pragma omp critical (IntegrationRemoveFrontier)
-//                  removed_frontier_blocks_->insert(block->code());
-//                }
-//              }
-//            }
-//          }
-//        }
-
         return;
       }
 
@@ -601,7 +464,6 @@ struct MultiresOFusionUpdate {
           // Update the voxel data based using the depth measurement
           const int voxel_idx = x + y * size_at_integration_scale_li + z * size_at_integration_scale_sq;
           auto& voxel_data = block->currData(voxel_idx); /// \note pass by reference now.
-          const VoxelState prev_state = voxel_data.state;
           if (low_variance) {
             block->incrCurrObservedCount(updating_model::freeVoxel(voxel_data));
           } else {
@@ -610,38 +472,11 @@ struct MultiresOFusionUpdate {
             const float range_diff = (sample_point_C_m - depth_value) * (range / sample_point_C_m);
             block->incrCurrObservedCount(updating_model::updateVoxel(range_diff, tau, three_sigma, rgba_value, fg_value, voxel_data));
           }
-          // Set the voxel state
-          if (added_frontier_blocks_ && removed_frontier_blocks_) {
-            if (MultiresOFusion::VoxelType::isFree(voxel_data)) {
-              const Eigen::Vector3i coord = voxel_coord_base + integration_stride * Eigen::Vector3i(x, y, z);
-              bool is_frontier;
-#pragma omp critical (IntegrationCheckFrontier)
-              is_frontier = se::isFrontier(coord, size_at_integration_scale_li, integration_scale, *block, map_);
-              if (is_frontier) {
-                voxel_data.state = VoxelState::Frontier;
-              } else {
-                voxel_data.state = VoxelState::Free;
-              }
-            } else if (MultiresOFusion::VoxelType::isInside(voxel_data)) {
-              voxel_data.state = VoxelState::Occupied;
-            } else {
-              voxel_data.state = VoxelState::Unknown;
-            }
-            // Update the frontier block sets
-            if (voxel_data.state == VoxelState::Frontier) {
-#pragma omp critical (IntegrationAddFrontier)
-              added_frontier_blocks_->insert(block->code());
-            } else if (prev_state == VoxelState::Frontier) {
-#pragma omp critical (IntegrationRemoveFrontier)
-              removed_frontier_blocks_->insert(block->code());
-            }
-          }
         } // x
       } // y
     } // z
 
     block->incrCurrIntegrCount();
-
   }
 
 
@@ -832,7 +667,7 @@ struct MultiresOFusionUpdate {
         auto& parent_data = block->parent()->childData(child_idx);
         parent_data = max_data;
 
-        if (   max_data.observed
+        if (   max_data.observed && !max_data.frontier
                && max_data.x * max_data.y <= 0.95 * MultiresOFusion::min_occupancy) {
           pool_.deleteBlock(block, voxel_depth_);
         }
@@ -860,7 +695,7 @@ struct MultiresOFusionUpdate {
           auto node_data = updating_model::propagateToNoteAtCoarserScale(node, voxel_depth_, frame_);
           node_list_[d-1].insert(node->parent());
 
-          if (   node_data.observed
+          if (   node_data.observed && !node_data.frontier
                  && node_data.x * node_data.y <= 0.95 * MultiresOFusion::min_occupancy) {
             pool_.deleteNode(node, voxel_depth_);
           }
@@ -1345,9 +1180,7 @@ void MultiresOFusion::integrate(OctreeType&             map,
                                 const Eigen::Matrix4f&  T_CM,
                                 const SensorImpl&       sensor,
                                 const unsigned          frame,
-                                std::set<se::key_t>*    free_nodes,
-                                std::set<se::key_t>*    added_frontier_blocks,
-                                std::set<se::key_t>*    removed_frontier_blocks) {
+                                std::set<se::key_t>*    updated_nodes) {
   TICKD("updateMap")
   // Create min/map depth pooling image for different bounding box sizes
   const std::unique_ptr<se::DensePoolingImage<SensorImpl>> pooling_depth_image(new se::DensePoolingImage<SensorImpl>(depth_image));
@@ -1363,8 +1196,7 @@ void MultiresOFusion::integrate(OctreeType&             map,
   std::vector<std::set<NodeType*>> node_list(map.blockDepth());
   MultiresOFusionUpdate<SensorImpl> funct(map, block_list, node_list, free_list, low_variance_list, projects_inside_list,
                                           depth_image, rgba_image, fg_image, pooling_depth_image.get(), sensor, T_CM,
-                                          voxel_dim, map.voxelDepth(), max_depth_value, frame,
-                                          free_nodes, added_frontier_blocks, removed_frontier_blocks);
+                                          voxel_dim, map.voxelDepth(), max_depth_value, frame);
 
   // Launch on the 8 voxels of the first depth
 #pragma omp parallel for
@@ -1376,6 +1208,13 @@ void MultiresOFusion::integrate(OctreeType&             map,
   }
 
   TOCK("mapToCameraAllocation")
+
+  // Everything in the block_list is a candidate frontier.
+  if (updated_nodes) {
+    for (const auto block : block_list) {
+      updated_nodes->insert(block->code());
+    }
+  }
 
   TICKD("updateBlock")
 #pragma omp parallel for
